split views to multiple files
