+++ mfgkanbanbackend/kanban/views/__init__.py
@@ -1,0 +1,3 @@
+from .part_views import PartCreate, PartDetail, PartList
+from .location_views import LocationCreate, LocationDetail, LocationList
+from .workorder_views import WorkOrderCreate, WorkOrderDetail, WorkOrderList
